Review comment from @somathias, rename the genomics-specific validation file to validate_genomics_schema.py
