+++ src/scilifelab_metadata_templates/genomics/README.md
@@ -4,7 +4,7 @@ This folder contains scripts for generating the SciLifeLab Genomics Technical Me
 - `generate_genomics_template.py`: main script to generate the genomics template and schema, see below for more information.
 - `data/technical_metadata_fields_incl_ENA_CVs.yml`: list of relevant technical metadata fields, including ENA controlled vocabularies where available
 - `update_ENA_controlled_vocabs.py`: functionality to fetch ENA's controlled vocabularies, based on [template_updater.py](https://github.com/ELIXIR-Belgium/ENA-metadata-templates/blob/main/scripts/template_updater.py), copyright @ Elixir Belgium and licensed under the MIT license.
-- `validate_json_schema.py`: script to validate filled-in tsv files against the genomics template json schema.
+- `validate_genomics_schema.py`: script to validate filled-in tsv files against the genomics template json schema.
 
 ## Regenerating the template
 ### Manually
@@ -34,14 +34,14 @@ The template files (_.tsv_, _.json_) are generated by the script [generate_genom
 3. It regenerates the json schema for the template. 
 4. It updates the list of the included metadata fields in the genomics readme.
 
-## Validate_json_schema.py
+## Validate_genomics_schema.py
 
-Filled-in _.tsv_ files can be validated against a genomics template json schema using the validation script [scripts/validate_json_schema.py](https://github.com/ScilifelabDataCentre/scilifelab-metadata-templates/blob/main/src/scilifelab_metadata_templates/genomics/validate_json_schema.py). It can be passed a multi-row _.tsv_ to validate as follows
+Filled-in _.tsv_ files can be validated against a genomics template json schema using the validation script [scripts/validate_genomics_schema.py](https://github.com/ScilifelabDataCentre/scilifelab-metadata-templates/blob/main/src/scilifelab_metadata_templates/genomics/validate_genomics_schema.py). It can be passed a multi-row _.tsv_ to validate as follows
 ```
-python validate_json_schema.py path/to/your/tsv/file.tsv
+python validate_genomics_schema.py path/to/your/tsv/file.tsv
 ```
 Optionally, the schema can be specified using 
 ```
-python validate_json_schema.py path/to/your.tsv --schema path/to/schema.json
+python validate_genomics_schema.py path/to/your.tsv --schema path/to/schema.json
 ```
 with the default being set as `../genomics_technical_metadata_schema.json`.
