+++ src/scilifelab_metadata_templates/genomics/__init__.py
@@ -1,3 +1,3 @@
-from .validate_json_schema import validate_genomics_data
+from .validate_genomics_schema import validate_genomics_data
 
 __all__ = ["validate_genomics_data"]
